refactoring: #845 rename module "formats" to "output_format"
