+++ osmaxx/conversion/converters/converter_gis/gis.py
@@ -15,7 +15,7 @@ from osmaxx.conversion._settings import odb_license
 from osmaxx.conversion.converters.converter_gis.bootstrap import BootStrapper
 from osmaxx.conversion.converters.converter_gis.extract.db_to_format.extract import extract_to
 from osmaxx.conversion.converters.utils import zip_folders_relative, recursive_getsize
-from osmaxx.conversion.constants.formats import FORMAT_DEFINITIONS
+from osmaxx.conversion.constants.output_format import FORMAT_DEFINITIONS
 
 
 def perform_export(
