+++ osmaxx/conversion/models.py
@@ -11,7 +11,7 @@ from osmaxx.clipping_area.models import ClippingArea
 from osmaxx.conversion.converters.converter import convert
 from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES, DETAIL_LEVEL_ALL
 from osmaxx.conversion.constants.coordinate_reference_systems import CRS_CHOICES
-from osmaxx.conversion.constants.formats import FORMAT_CHOICES
+from osmaxx.conversion.constants.output_format import FORMAT_CHOICES
 from osmaxx.conversion.constants.statuses import STATUS_CHOICES, RECEIVED
