+++ osmaxx/conversion/size_estimator.py
@@ -2,35 +2,35 @@ import math
 
 from osmaxx.conversion.converters.converter_gis import detail_levels
 from osmaxx.conversion.models import Job
-from osmaxx.conversion.constants import formats
+from osmaxx.conversion.constants import output_format
 
 PRE_DATA = {
-    formats.GARMIN: {
+    output_format.GARMIN: {
         'pbf_predicted': [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_ALL: [11000, 18000, 42000, 95000],
         detail_levels.DETAIL_LEVEL_REDUCED: [11000, 18000, 42000, 95000],
     },
-    formats.PBF: {
+    output_format.PBF: {
         'pbf_predicted': [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_ALL: [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_REDUCED: [25000, 44000, 96000, 390000],
     },
-    formats.FGDB: {
+    output_format.FGDB: {
         'pbf_predicted': [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_ALL: [46000, 101000, 309000, 676000],
         detail_levels.DETAIL_LEVEL_REDUCED: [21000, 27000, 107000, 250000],
     },
-    formats.GPKG: {
+    output_format.GPKG: {
         'pbf_predicted': [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_ALL: [109000, 210000, 690000, 1500000],
         detail_levels.DETAIL_LEVEL_REDUCED: [49000, 58000, 252000, 599000],
     },
-    formats.SHAPEFILE: {
+    output_format.SHAPEFILE: {
         'pbf_predicted': [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_ALL: [255000, 638000, 2000000, 4400000],
         detail_levels.DETAIL_LEVEL_REDUCED: [100000, 138000, 652000, 1600000],
     },
-    formats.SPATIALITE: {
+    output_format.SPATIALITE: {
         'pbf_predicted': [25000, 44000, 96000, 390000],
         detail_levels.DETAIL_LEVEL_ALL: [115000, 216000, 719000, 1600000],
         detail_levels.DETAIL_LEVEL_REDUCED: [55000, 66000, 269000, 635000],
@@ -49,7 +49,7 @@ def size_estimation_for_format(format_type, detail_level, predicted_pbf_size):
 
 
 def get_data(format_type, detail_level):
-    assert format_type in formats.FORMAT_DEFINITIONS
+    assert format_type in output_format.FORMAT_DEFINITIONS
     assert detail_level in [level[0] for level in detail_levels.DETAIL_LEVEL_CHOICES]
     base_query_set = Job.objects.filter(
         parametrization__out_format=format_type,
