+++ tests/excerptexport/conftest.py
@@ -4,7 +4,7 @@ from collections import namedtuple
 
 import pytest
 
-from osmaxx.conversion.constants import formats
+from osmaxx.conversion.constants import output_format
 from osmaxx.excerptexport.models import Excerpt, ExtractionOrder, Export, OutputFile
 
 
@@ -29,7 +29,7 @@ def extraction_order(excerpt, user, db):
     return extraction_order
 
 
-@pytest.fixture(params=formats.ALL)
+@pytest.fixture(params=output_format.ALL)
 def export(request, extraction_order):
     return Export.objects.create(
         extraction_order=extraction_order,
@@ -40,7 +40,7 @@ def export(request, extraction_order):
 @pytest.fixture
 def exports(extraction_order):
     ParamFake = namedtuple('ParamFake', 'param')
-    return [export(ParamFake(format), extraction_order) for format in formats.ALL]
+    return [export(ParamFake(format), extraction_order) for format in output_format.ALL]
 
 
 @pytest.fixture
