+++ tests/excerptexport/models/test_export_model.py
@@ -23,10 +23,10 @@ def export_stale_status(mocker, export):
 @pytest.mark.django_db
 def test_created_export_has_correct_time_stamp(extraction_order):
     from osmaxx.excerptexport.models import Export
-    from osmaxx.conversion.constants import formats
+    from osmaxx.conversion.constants import output_format
     export = Export.objects.create(
         extraction_order=extraction_order,
-        file_format=formats.FGDB,
+        file_format=output_format.FGDB,
     )
     now = timezone.now()
     margin = timedelta(minutes=1)
